Add the anachronisms.h header.
Recognizing that this repo isn't really an actively developed project
but instead a historical snapshot of code from more then 30 years ago,
the question becomes: How can we make the code compile so that it can
be experienced on a modern system while minimizing the alterations to
the code itself?  My conclusion was that a header composed entirely
anachronistic changes was a good start.

This commit eliminates nearly 1000 lines of warnings simply by
including all of the required headers at the beginning of most files.
The one change that wasn't simply adding an include statement was a
required change in bprintf.c.  The prototype for malloc() was just
wrong, and caused errors when the correct header was included.

I marked it with an "anachronism" comment, which I intend to continue
using to mark all places that are changed.