Eliminate warnings caused by void/int return value confusion
Having just said that the goal was to minimize changes to the original
source, I was very conflicted about this commit.  The problem is that
the warning:

warning: ‘return’ with no value, in function returning non-void

has no easy option to disable it, and there's a couple thousand lines
worth of these errors.  Originally I thought that I would fix these,
but then not commit them so that I could minimize changes to the
source.  That turned out to be complicated by the commit that follows,
and eventually I realized that the benefits of having a clean compile
going forward probably outweighed the size of the changes to the
source, expecially since this commit is basically 216 indentical added
lines.  It's not like it's hard to recognize these, or understand what
changed.

Even still, I minimized changes to the original source by only
correcting functions that caused warnings, but not those that returned
void in practice but didn't actually cause a warning.  Any that were
marked that didn't generate a warning were probably something that
ended up needing to be void for some otehr reason later on.