Switch from suppressing all errors to just -Wno-implicit-int.
warning: return type defaults to ‘int’ [-Wimplicit-int]
warning: type of ‘x’ defaults to ‘int’ [-Wimplicit-int]

Both of these defaults seem to have been well understood by the
original authors, and while I'd prefer to be explicit in any modern
code, it doesn't seem like these are likely to be the cause of the
crashes, and they can be easily disabled with a compiler option.  This
reduces the warnings by approximately 50% (another 1000 lines), and
brings the total number of remaining warnings down to about 500.