Fix prototypes for signal handlers in gamego.c.
This fix was a little more intrusive than I would have liked.  I
thought I could fix it by redefining signal in anachronisms.h, but
that turned out to be more complicated than that because of the
incorrect forward declarations and the use of SIG_IGN, which now has
the correct definfition from the header. :-(

In the end it seemed best to just fix this even thought I doubt it's
the source of any serious problems.