Format string warnings, a couple of which are definitely wrong.
Most of these were the relatively harmless distinction between %d and
%ld like sthis:

warning: format ‘%d’ expects argument of type ‘int’, but
  argument 3 has type ‘long int’ [-Wformat=]

But there were a couple of more serious errors as well.  The scanf()
may or may not be doing the right thing with mismatched types,
printing a string pointer with %d will obviously result in nonsense.