Convert bprintf to be a proper varargs function.
The existing implementation was a recipe for disaster given the
hardcoded types and sizes of the arguments.  I considered less
intrusive solutions, but in the end they weren't really any better
because of the complications involving the bogus delcaration in
System.h and bprintf's depencies on quprnt() and syslog.

Regarding the stub in gmain2.c...  Aren't compilers supposed to tell
you whether a function is being called? ;-)