Convert syslog() to a proper varargs function.
Same situation as bprintf().  Several (~5) lines of fixes in one place
(at syslog) weighed against correcting the inconsistent number of
arguments in four places, plus the rather more serious problem of
using %d and providing a parameter that was not in fact a pointer at
all.  Besides this approach just feels cleaner.