Fix warnings about the ambiguously typed last parameter of sendsys.
Even after the previous two commits resolved 15 places where a long*
was passed as the last argument to sendsys(), where a char* was
obviously intended, there are three locations seem to be legimately
passing a long*.  Fortunately, these are the cases where codeword is
-9900 or -10021, so it seems that this was known to the author, and
possibly led to the ambiguity about how to type this parameter.

Pointers are all the same size, so I don't think is a crashing issue,
but to clean up the ambiguity somehow.  It seems cleaner (and less
intrusive) to fix it here which a void* type, and then cast it as char
and long within this function than to type cast it in the three places
where sendsys() is called.