Fix a crashing problem in getreinput().
Because the parameter defaulted to type "int" but was deferenced as a
pointer, and pointers on 64-bit systems are twice as large, there was
a parameter size mismatch and a crash.  Correctly typing the parameter
fixed the problem.  Sadly, I expect that there are more of this class
of bug since -Wno-implict-int is still enabled, but changing that is a
pretty intrusive change. :-(