Fix confusion about rte() parameter count.
So this bug was a little troublesome, since it required "inventing" a
parameter where one was not supplied despite the function _requiring_
a parameter.  There was no easy fix here as the original code was just
wrong.  My conclusion, which was based on nothing more really than an
impression of what the code does, was that the function used to assume
globme, but when it was modified to support a parameter for calls in
tk.c, a few of the old calls were missed.

So I passed globme.  No idea if that's correct or not, but it should
prevent the crash, and it puts something in the value pointed to that
should be "valid" semantically based on lots of other calls to rte().

I'm very surprised that this didn't crash in the 32-bit builds as
well.  There's probably something "rare" and/or untested here as well.