Cleanup various delu2 problems.
This function is frankly a mess.  There's an implementation in
gmain2.c that's pretty rational, except for lacking a forward
declaration.

There's another implemenation in ndebug that appears to be nothing but
a stub, and yet's it's being called from magic.c with the wrong API.
The call expects a return code with neither implementation actually
provides, and passes a parameter that the version it links against
doesn't actually accept.  Previously I unified these implementations
into one declaration of the form "void delu2(char*)", but on further
consideration, it's clear that the intention of the code is also
muddled.  Perhaps the call should be removed entirely, or even linked
against the more complete implementation in gmain2.c

This solution is attractive mostly in minimizing changes to the code.