Revert "Suppress some %ld versus %d warnings."
This reverts commit 5fa8d2ca41a2991b0286070ec532cebb2ef2f774.

And amend it to use the %ld's after all.  I had missed that the scanf
foirmat string would need to change as well, and while the former fix
would have affected two lines, I think this version that affects four
lines is closer to the original code's intention even if it is more
intrusive.  The deciding factor was that this code is more correct
when compiled on 32-bit systems.