Fix warning about implicit declarations for chpwd().
Previously I had fixed this by declaring in anachronisms.h, but the
case is very similar that of delu2() with regards to the interaction
with ndebug.c.  This commit chooses consistency with the delu2()
approach by forward declaring in gmain2.c and magic.c, with the
benefit of removing a change from ndebug.

As with delu2() the difference in API between the two compilation
units, but hopefully this approach helps draw attention to that rather
then covering it over more.