A different approach for getpass() and getpw().
The changes to blib.c have NOT been marked as anachronisms, because
they've actually reverting changes made since the DOS ZIP archive.  I
had originally hoped to fix more things in the anachronisms header,
but it turned out that most of the problems weren't easily fixed this
way.

Obviously the defines are _very_ specific to the places where the APIs
are being called, but I think that's ok, since they're only called
from the one spot, and we're nopt planning on "maintaining" the code.