A little CMakefiles.tx cleanup.
At this point, the following commands:

mkdir build && cd build && cmake .. && make && make reset && ./mud.1

result in an AberMud that makes it father than it used to, and it
compiles cleanly with only one warning, and I'm fairly certain that's
not a warning but a longish chain of type failures that cause crashes.

For testing: make clean && make && make reset && ./mud.1

Works pretty reliably (at least for me).