Fix embarrassing mistake on my part.
So the "v" in front of snprintf and fprintf in my bprintf and syslog
reimplementations is NOT optional. ;-) On the positive side, with this
change output is working again, and the mud is looking more functional
than it ever has before (for me).  I even saw a mob after a reset! ;-)