Suppress the last warning with a hacky solution.
This relies on the fact the long and FILE* are coincidentally the same
size.  I'd prefer not to fix this bug this way, but it actually turns
into a quite intrusive change to correctly type all FILE pointers.
Since this seems to be working and minimizing differences is a goal,
it will get the job done.